+++ packages/subtree-updater/src/server.ts
@@ -19,6 +19,7 @@ export class SubtreeUpdateServer {
   private stopped: boolean;
   private prom?: Promise<void>;
   private timer?: NodeJS.Timeout;
+  private fillBatches: boolean;
 
   constructor(
     prover: SubtreeUpdateProver,
@@ -53,27 +54,36 @@ export class SubtreeUpdateServer {
   public async start(): Promise<void> {
     this.stopped = false;
     const prom = new Promise<void>((resolve, reject) => {
-      this.timer = setInterval(async () => {
+      const poll = async () => {
         if (this.stopped) {
           resolve(undefined);
           return;
         }
 
         try {
+          if (this.fillBatches) {
+          }
+
           console.log("polling for batch...");
           const filledBatch =
             await this.updater.pollInsertionsAndTryMakeBatch();
 
           if (filledBatch) {
-            console.log("filled batch! generating and submitting proof");
+            console.log("filled batch!");
+            console.log("generating and submitting proof...");
             await this.updater.tryGenAndSubmitProofs();
-            console.log("proof submitted");
+            console.log("proof submitted!");
+          } else if (this.fillBatches) {
+            console.log("batch not yet full. filling it with zeros...");
+            await this.updater.fillBatch();
+            await poll();      
           }
         } catch (err) {
           console.error("subtree update server received an error:", err);
           reject(err);
         }
-      }, this.interval);
+      }
+      this.timer = setInterval(poll, this.interval);
     });
 
     this.prom = prom.finally(() => clearTimeout(this.timer));